refactor: move interface to relevant file
also, replace string w/ specific type
