+++ packages/sdk/src/transactions/types.ts
@@ -54,3 +54,18 @@ export interface ScriptPubKey {
   address: string;
   type: string;
 }
+
+export interface UTXO {
+  n: number;
+  txHash: string;
+  blockHash: string;
+  blockN: number;
+  sats: number;
+  scriptPubKey: ScriptPubKey;
+  txid: string;
+  value: number;
+  ordinals?: Ordinal[] | null;
+  inscriptions?: Inscription[] | null;
+  safeToSpend: boolean;
+  confirmation: number;
+}
